microservices/products-service/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
requests==2.31.0

